Refact dev mode
