+++ node/parallel/Cargo.toml
@@ -25,15 +25,16 @@ async-trait            = '0.1.42'
 
 primitives = { package = 'parallel-primitives', path = '../../primitives' }
 
-heiko-runtime                              = { path = '../../runtime/heiko' }
-kerria-runtime                             = { path = '../../runtime/kerria' }
 pallet-loans-rpc                           = { path = '../../pallets/loans/rpc' }
 pallet-router-rpc                          = { path = '../../pallets/router/rpc' }
 pallet-transaction-payment-rpc             = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.26' }
 pallet-transaction-payment-rpc-runtime-api = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.26' }
+
+# runtimes
+heiko-runtime                              = { path = '../../runtime/heiko' }
 parallel-runtime                           = { path = '../../runtime/parallel' }
-vanilla-runtime                            = { path = '../../runtime/vanilla'  }
-dev-runtime                                = { path = '../../runtime/dev', optional = true }
+kerria-runtime                             = { path = '../../runtime/kerria', optional = true}
+vanilla-runtime                            = { path = '../../runtime/vanilla', optional = true }
 
 sc-basic-authorship     = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.26' }
 sc-chain-spec           = { git = 'https://github.com/paritytech/substrate.git', branch = 'polkadot-v0.9.26' }
@@ -126,20 +127,19 @@ default            = ['std']
 runtime-benchmarks = [
   'parallel-runtime/runtime-benchmarks',
   'heiko-runtime/runtime-benchmarks',
-  'vanilla-runtime/runtime-benchmarks',
   'kerria-runtime/runtime-benchmarks',
   'polkadot-service/runtime-benchmarks',
+  "try-runtime-cli",
 ]
 std                = []
 try-runtime        = [
   'heiko-runtime/try-runtime',
   'parallel-runtime/try-runtime',
-  'vanilla-runtime/try-runtime',
   'kerria-runtime/try-runtime',
-  'try-runtime-cli',
+  "polkadot-service/try-runtime",
 ]
-with-evm-dev-runtime = [
-  'dev-runtime',
-  'dev-runtime/runtime-benchmarks',
-  'dev-runtime/try-runtime',
+with-evm-runtime = [
+  'vanilla-runtime',
+  'vanilla-runtime/runtime-benchmarks',
+  'vanilla-runtime/try-runtime',
 ]
