+++ node/parallel/src/command.rs
@@ -40,8 +40,8 @@ const HEIKO_PARA_ID: u32 = 2085;
 
 fn load_spec(id: &str) -> std::result::Result<Box<dyn sc_service::ChainSpec>, String> {
     Ok(match id {
-        #[cfg(feature = "with-evm-dev-runtime")]
-        "dev" => Box::new(chain_spec::dev::development_config(ParaId::from(
+        #[cfg(feature = "with-evm-runtime")]
+        "dev" => Box::new(chain_spec::vanilla::local_development_config(ParaId::from(
             HEIKO_PARA_ID,
         ))),
         "heiko-dev" => Box::new(chain_spec::heiko::heiko_dev_config(ParaId::from(
@@ -56,6 +56,7 @@ fn load_spec(id: &str) -> std::result::Result<Box<dyn sc_service::ChainSpec>, St
         "" | "parallel" => Box::new(chain_spec::parallel::parallel_config(ParaId::from(
             PARALLEL_PARA_ID,
         ))?),
+        #[cfg(feature = "with-evm-runtime")]
         "vanilla-dev" => Box::new(chain_spec::vanilla::vanilla_dev_config(ParaId::from(
             HEIKO_PARA_ID,
         ))),
@@ -76,7 +77,15 @@ fn load_spec(id: &str) -> std::result::Result<Box<dyn sc_service::ChainSpec>, St
             } else if starts_with("heiko") {
                 Box::new(chain_spec::heiko::ChainSpec::from_json_file(path)?)
             } else if starts_with("vanilla") {
-                Box::new(chain_spec::vanilla::ChainSpec::from_json_file(path)?)
+                #[cfg(feature = "with-evm-runtime")]
+                {
+                    Box::new(chain_spec::vanilla::ChainSpec::from_json_file(path)?)
+                }
+                #[cfg(not(feature = "with-evm-runtime"))]
+                return Err(
+                    "chain_spec's filename start with vanilla should be built with evm runtime"
+                        .into(),
+                );
             } else if starts_with("kerria") {
                 Box::new(chain_spec::kerria::ChainSpec::from_json_file(path)?)
             } else {
@@ -122,12 +131,13 @@ impl SubstrateCli for Cli {
             &parallel_runtime::VERSION
         } else if chain_spec.is_heiko() {
             &heiko_runtime::VERSION
-        } else if chain_spec.is_vanilla() {
-            &vanilla_runtime::VERSION
+        } else if chain_spec.is_vanilla() || chain_spec.is_dev() {
+            #[cfg(feature = "with-evm-runtime")]
+            return &vanilla_runtime::VERSION;
+            #[cfg(not(feature = "with-evm-runtime"))]
+            return &heiko_runtime::VERSION;
         } else if chain_spec.is_kerria() {
             &kerria_runtime::VERSION
-        } else if chain_spec.is_dev() {
-            &vanilla_runtime::VERSION
         } else {
             unreachable!()
         }
@@ -177,48 +187,53 @@ impl SubstrateCli for RelayChainCli {
 macro_rules! switch_runtime {
     ($chain_spec:expr, { $( $code:tt )* }) => {
         if $chain_spec.is_parallel() {
-			#[allow(unused_imports)]
-            use crate::service::ParallelExecutor as Executor;
-			#[allow(unused_imports)]
-            use parallel_runtime::{RuntimeApi, Block};
+            #[cfg(not(feature = "with-evm-runtime"))] {
+                #[allow(unused_imports)]
+                use crate::service::ParallelExecutor as Executor;
+                #[allow(unused_imports)]
+                use parallel_runtime::{RuntimeApi, Block};
 
-			$( $code )*
-        } else if $chain_spec.is_heiko() {
-			#[allow(unused_imports)]
-            use crate::service::HeikoExecutor as Executor;
-			#[allow(unused_imports)]
-            use heiko_runtime::{RuntimeApi, Block};
-
-			$( $code )*
-        } else if $chain_spec.is_vanilla() {
-			#[allow(unused_imports)]
-            use crate::service::VanillaExecutor as Executor;
-			#[allow(unused_imports)]
-            use vanilla_runtime::{RuntimeApi, Block};
-
-			$( $code )*
-        } else if $chain_spec.is_kerria() {
-			#[allow(unused_imports)]
-            use crate::service::KerriaExecutor as Executor;
-			#[allow(unused_imports)]
-            use kerria_runtime::{RuntimeApi, Block};
-
-			$( $code )*
-        } else if $chain_spec.is_dev() {
-            #[cfg(feature = "with-evm-dev-runtime")]
-            #[allow(unused_imports)]
-            use crate::dev::service::Executor as Executor;
-            #[cfg(feature = "with-evm-dev-runtime")]
-            #[allow(unused_imports)]
-            use dev_runtime::{RuntimeApi, Block};
-            #[cfg(feature = "with-evm-dev-runtime")]
-            {
                 $( $code )*
             }
-            #[cfg(not(feature = "with-evm-dev-runtime"))]
-			return Err("compile the node with `--features with-evm-dev-runtime`".into());
-        }
-        else {
+            #[cfg(feature = "with-evm-runtime")] {
+                unreachable!();
+            }
+        } else if $chain_spec.is_heiko() {
+            #[cfg(not(feature = "with-evm-runtime"))] {
+                #[allow(unused_imports)]
+                use crate::service::HeikoExecutor as Executor;
+                #[allow(unused_imports)]
+                use heiko_runtime::{RuntimeApi, Block};
+
+                $( $code )*
+            }
+            #[cfg(feature = "with-evm-runtime")] {
+                unreachable!();
+            }
+        } else if ($chain_spec.is_vanilla() || $chain_spec.is_dev()) {
+            #[cfg(feature = "with-evm-runtime")] {
+                #[allow(unused_imports)]
+                use crate::evm_service::VanillaExecutor as Executor;
+                #[allow(unused_imports)]
+                use vanilla_runtime::{RuntimeApi, Block};
+
+                $( $code )*
+            }#[cfg(not(feature = "with-evm-runtime"))] {
+                unreachable!();
+            }
+        } else if $chain_spec.is_kerria() {
+            #[cfg(not(feature = "with-evm-runtime"))] {
+                #[allow(unused_imports)]
+                use crate::service::KerriaExecutor as Executor;
+                #[allow(unused_imports)]
+                use kerria_runtime::{RuntimeApi, Block};
+
+                $( $code )*
+            }
+            #[cfg(feature = "with-evm-runtime")] {
+                unreachable!();
+            }
+        } else {
             unreachable!();
         }
     };
@@ -434,10 +449,10 @@ pub fn run() -> Result<()> {
 
             switch_runtime!(chain_spec, {
                 runner.run_node_until_exit(|config| async move {
-                    #[cfg(feature = "with-evm-dev-runtime")]
+                    #[cfg(feature = "with-evm-runtime")]
                     {
                         if config.chain_spec.is_dev() {
-                            return crate::dev::start_node(config).map_err(Into::into);
+                            return crate::evm_service::start_dev_node::<RuntimeApi, Executor>(config).map_err(Into::into);
                         }
                     }
 
@@ -491,12 +506,15 @@ pub fn run() -> Result<()> {
                         }
                     );
 
-                    if config.chain_spec.is_vanilla() {
-                        crate::evm_service::start_node(config, polkadot_config, collator_options, id)
+                    #[cfg(feature = "with-evm-runtime")]
+                    {
+                        crate::evm_service::start_node::<RuntimeApi, Executor>(config, polkadot_config, collator_options, id)
                             .await
                             .map(|r| r.0)
                             .map_err(Into::into)
-                    } else{
+                    }
+                    #[cfg(not(feature = "with-evm-runtime"))]
+                    {
                         crate::service::start_node::<RuntimeApi, Executor>(
                             config,
                             polkadot_config,
