--- node/parallel/src/dev/service.rs
@@ -1,344 +1,0 @@
-// Copyright 2021 Parallel Finance Developer.
-// This file is part of Parallel Finance.
-
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-use cumulus_primitives_parachain_inherent::MockValidationDataInherentDataProvider;
-#[allow(dead_code, unused)]
-use fc_consensus::FrontierBlockImport;
-use fc_rpc_core::types::{FeeHistoryCache, FilterPool};
-use futures::StreamExt;
-use sc_client_api::BlockchainEvents;
-use sc_consensus_manual_seal::{self as manual_seal};
-use sc_executor::NativeElseWasmExecutor;
-use sc_service::{error::Error as ServiceError, Configuration, TaskManager};
-use sc_telemetry::{Telemetry, TelemetryWorker};
-use sp_blockchain::HeaderBackend;
-use std::{collections::BTreeMap, sync::Arc, time::Duration};
-
-pub use dev_runtime::{self, RuntimeApi, SLOT_DURATION};
-
-use primitives::*;
-
-/// Local runtime native executor.
-pub struct Executor;
-
-impl sc_executor::NativeExecutionDispatch for Executor {
-    #[cfg(not(feature = "runtime-benchmarks"))]
-    type ExtendHostFunctions = ();
-
-    #[cfg(feature = "runtime-benchmarks")]
-    type ExtendHostFunctions = frame_benchmarking::benchmarking::HostFunctions;
-
-    fn dispatch(method: &str, data: &[u8]) -> Option<Vec<u8>> {
-        dev_runtime::api::dispatch(method, data)
-    }
-
-    fn native_version() -> sc_executor::NativeVersion {
-        dev_runtime::native_version()
-    }
-}
-
-type FullClient = sc_service::TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>;
-type FullBackend = sc_service::TFullBackend<Block>;
-type FullSelectChain = sc_consensus::LongestChain<FullBackend, Block>;
-
-/// Build a partial chain component config
-pub fn new_partial(
-    config: &Configuration,
-) -> Result<
-    sc_service::PartialComponents<
-        FullClient,
-        FullBackend,
-        FullSelectChain,
-        sc_consensus::DefaultImportQueue<Block, FullClient>,
-        sc_transaction_pool::FullPool<Block, FullClient>,
-        (
-            FrontierBlockImport<Block, Arc<FullClient>, FullClient>,
-            Option<Telemetry>,
-            Arc<fc_db::Backend<Block>>,
-        ),
-    >,
-    ServiceError,
-> {
-    if config.keystore_remote.is_some() {
-        return Err(ServiceError::Other(
-            "Remote Keystores are not supported.".to_string(),
-        ));
-    }
-
-    let telemetry = config
-        .telemetry_endpoints
-        .clone()
-        .filter(|x| !x.is_empty())
-        .map(|endpoints| -> Result<_, sc_telemetry::Error> {
-            let worker = TelemetryWorker::new(16)?;
-            let telemetry = worker.handle().new_telemetry(endpoints);
-            Ok((worker, telemetry))
-        })
-        .transpose()?;
-    let executor = sc_executor::NativeElseWasmExecutor::<Executor>::new(
-        config.wasm_method,
-        config.default_heap_pages,
-        config.max_runtime_instances,
-        config.runtime_cache_size,
-    );
-
-    let (client, backend, keystore_container, task_manager) =
-        sc_service::new_full_parts::<Block, RuntimeApi, _>(
-            config,
-            telemetry.as_ref().map(|(_, telemetry)| telemetry.handle()),
-            executor,
-        )?;
-    let client = Arc::new(client);
-    let telemetry = telemetry.map(|(worker, telemetry)| {
-        task_manager
-            .spawn_handle()
-            .spawn("telemetry", None, worker.run());
-        telemetry
-    });
-    let select_chain = sc_consensus::LongestChain::new(backend.clone());
-    let transaction_pool = sc_transaction_pool::BasicPool::new_full(
-        config.transaction_pool.clone(),
-        config.role.is_authority().into(),
-        config.prometheus_registry(),
-        task_manager.spawn_essential_handle(),
-        client.clone(),
-    );
-    let frontier_backend = crate::evm_rpc::open_frontier_backend(config)?;
-
-    let frontier_block_import =
-        FrontierBlockImport::new(client.clone(), client.clone(), frontier_backend.clone());
-
-    let import_queue = sc_consensus_manual_seal::import_queue(
-        Box::new(frontier_block_import.clone()),
-        &task_manager.spawn_essential_handle(),
-        config.prometheus_registry(),
-    );
-
-    Ok(sc_service::PartialComponents {
-        client,
-        backend,
-        task_manager,
-        import_queue,
-        keystore_container,
-        select_chain,
-        transaction_pool,
-        other: (frontier_block_import, telemetry, frontier_backend),
-    })
-}
-
-/// Builds a new service.
-pub fn start_node(config: Configuration) -> Result<TaskManager, ServiceError> {
-    let sc_service::PartialComponents {
-        client,
-        backend,
-        mut task_manager,
-        import_queue,
-        keystore_container,
-        select_chain,
-        transaction_pool,
-        other: (block_import, mut telemetry, frontier_backend),
-    } = new_partial(&config)?;
-
-    let (network, system_rpc_tx, network_starter) =
-        sc_service::build_network(sc_service::BuildNetworkParams {
-            config: &config,
-            client: client.clone(),
-            transaction_pool: transaction_pool.clone(),
-            spawn_handle: task_manager.spawn_handle(),
-            import_queue,
-            block_announce_validator_builder: None,
-            warp_sync: None,
-        })?;
-
-    if config.offchain_worker.enabled {
-        sc_service::build_offchain_workers(
-            &config,
-            task_manager.spawn_handle(),
-            client.clone(),
-            network.clone(),
-        );
-    }
-
-    let filter_pool: FilterPool = Arc::new(std::sync::Mutex::new(BTreeMap::new()));
-    let fee_history_cache: FeeHistoryCache = Arc::new(std::sync::Mutex::new(BTreeMap::new()));
-    let overrides = crate::evm_rpc::overrides_handle(client.clone());
-
-    // Frontier offchain DB task. Essential.
-    // Maps emulated ethereum data to substrate native data.
-    task_manager.spawn_essential_handle().spawn(
-        "frontier-mapping-sync-worker",
-        Some("frontier"),
-        fc_mapping_sync::MappingSyncWorker::new(
-            client.import_notification_stream(),
-            Duration::new(6, 0),
-            client.clone(),
-            backend.clone(),
-            frontier_backend.clone(),
-            3,
-            0,
-            fc_mapping_sync::SyncStrategy::Parachain,
-        )
-        .for_each(|()| futures::future::ready(())),
-    );
-
-    // Frontier `EthFilterApi` maintenance. Manages the pool of user-created Filters.
-    // Each filter is allowed to stay in the pool for 100 blocks.
-    const FILTER_RETAIN_THRESHOLD: u64 = 100;
-    task_manager.spawn_essential_handle().spawn(
-        "frontier-filter-pool",
-        Some("frontier"),
-        fc_rpc::EthTask::filter_pool_task(
-            client.clone(),
-            filter_pool.clone(),
-            FILTER_RETAIN_THRESHOLD,
-        ),
-    );
-
-    const FEE_HISTORY_LIMIT: u64 = 2048;
-    task_manager.spawn_essential_handle().spawn(
-        "frontier-fee-history",
-        Some("frontier"),
-        fc_rpc::EthTask::fee_history_task(
-            client.clone(),
-            overrides.clone(),
-            fee_history_cache.clone(),
-            FEE_HISTORY_LIMIT,
-        ),
-    );
-
-    let role = config.role.clone();
-    let prometheus_registry = config.prometheus_registry().cloned();
-    let is_authority = config.role.is_authority();
-
-    let block_data_cache = Arc::new(fc_rpc::EthBlockDataCacheTask::new(
-        task_manager.spawn_handle(),
-        overrides.clone(),
-        50,
-        50,
-        prometheus_registry.clone(),
-    ));
-
-    let rpc_extensions_builder = {
-        let client = client.clone();
-        let network = network.clone();
-        let transaction_pool = transaction_pool.clone();
-
-        Box::new(move |deny_unsafe, subscription| {
-            let deps = crate::evm_rpc::FullDeps {
-                client: client.clone(),
-                pool: transaction_pool.clone(),
-                graph: transaction_pool.pool().clone(),
-                network: network.clone(),
-                is_authority,
-                deny_unsafe,
-                frontier_backend: frontier_backend.clone(),
-                filter_pool: filter_pool.clone(),
-                fee_history_limit: FEE_HISTORY_LIMIT,
-                fee_history_cache: fee_history_cache.clone(),
-                block_data_cache: block_data_cache.clone(),
-                overrides: overrides.clone(),
-            };
-
-            let io = crate::evm_rpc::create_full(deps, subscription)
-                .map_err::<ServiceError, _>(Into::into)?;
-
-            Ok(io)
-        })
-    };
-
-    let _rpc_handlers = sc_service::spawn_tasks(sc_service::SpawnTasksParams {
-        network: network.clone(),
-        client: client.clone(),
-        keystore: keystore_container.sync_keystore(),
-        task_manager: &mut task_manager,
-        transaction_pool: transaction_pool.clone(),
-        rpc_builder: rpc_extensions_builder,
-        backend,
-        system_rpc_tx,
-        config,
-        telemetry: telemetry.as_mut(),
-    })?;
-
-    if role.is_authority() {
-        let env = sc_basic_authorship::ProposerFactory::new(
-            task_manager.spawn_handle(),
-            client.clone(),
-            transaction_pool.clone(),
-            prometheus_registry.as_ref(),
-            telemetry.as_ref().map(|x| x.handle()),
-        );
-
-        let target_gas_price = 1;
-        let commands_stream = transaction_pool
-            .pool()
-            .clone()
-            .validated_pool()
-            .import_notification_stream()
-            .map(
-                |_| sc_consensus_manual_seal::rpc::EngineCommand::SealNewBlock {
-                    create_empty: true,
-                    finalize: true,
-                    parent_hash: None,
-                    sender: None,
-                },
-            );
-        let client_for_cidp = client.clone();
-
-        // Background authorship future
-        let authorship_future = manual_seal::run_manual_seal(manual_seal::ManualSealParams {
-            block_import,
-            env,
-            client: client.clone(),
-            pool: transaction_pool.clone(),
-            commands_stream,
-            select_chain,
-            consensus_data_provider: None,
-            create_inherent_data_providers: move |block: Hash, _| {
-                let current_para_block = client_for_cidp
-                    .number(block)
-                    .expect("Header lookup should succeed")
-                    .expect("Header passed in as parent should be present in backend.");
-
-                async move {
-                    let dynamic_fee =
-                        fp_dynamic_fee::InherentDataProvider(sp_core::U256::from(target_gas_price));
-
-                    let mocked_parachain = MockValidationDataInherentDataProvider {
-                        current_para_block,
-                        relay_offset: 1000,
-                        relay_blocks_per_para_block: 2,
-                        xcm_config: Default::default(),
-                        raw_downward_messages: vec![],
-                        raw_horizontal_messages: vec![],
-                    };
-
-                    Ok((
-                        sp_timestamp::InherentDataProvider::from_system_time(),
-                        mocked_parachain,
-                        dynamic_fee,
-                    ))
-                }
-            },
-        });
-        // we spawn the future on a background thread managed by service.
-        task_manager.spawn_essential_handle().spawn_blocking(
-            "instant-seal",
-            None,
-            authorship_future,
-        );
-    }
-    log::info!("Manual Seal Ready");
-
-    network_starter.start_network();
-    Ok(task_manager)
-}
